feat: 金额改为整数分存储与计算（int64 分），存量库启动迁移
后端：
- 所有金额字段由 float64（元）改为 int64（分）：采购总金额/单价、
  库存单价、结算回收额、回收明细扣减、摊销基数/日摊销、审计影响金额
- 摊销舍入：每日摊销 = 基数×当日执行次数÷计划次数（向下取整），
  尾差归最后一个执行日，保证 Σ(每日摊销) == round(基数×总执行÷计划)
- 结算唯一索引：一活动最多一条 settled 结算记录（并发兜底）
- 存量库迁移 migrateAmountToCents：PRAGMA user_version 0→1 幂等，
  一次性 ROUND(col*100)；main_test.go / settlement handler 测试

前端：
- 展示统一 /100 转元，采购总金额输入仍为元、提交 ×100 转分
- types / settlement / material / report 各页金额渲染适配

CLAUDE.md 记录设计决策（[[amount-in-cents]]）

+++ server/internal/model/settlement.go
@@ -8,7 +8,7 @@ type Settlement struct {
 	ActivityID          uint      `gorm:"not null;index" json:"activity_id"`
 	Status              string    `gorm:"size:20;not null;default:'settled'" json:"status"` // settled=已结算 reversed=已回撤
 	OperatorID          uint      `gorm:"not null" json:"operator_id"`
-	TotalReturnedAmount float64   `gorm:"not null" json:"total_returned_amount"` // 本次回收总金额
+	TotalReturnedAmount int64     `gorm:"not null" json:"total_returned_amount"` // 本次回收总金额（单位：分）
 	CreatedAt           time.Time `json:"created_at"`
 }
 
@@ -25,7 +25,7 @@ type RecoveryItem struct {
 	SettlementID  uint      `gorm:"not null;index" json:"settlement_id"`
 	StockID       uint      `gorm:"not null" json:"stock_id"`        // 回收物资ID
 	Quantity      int       `gorm:"not null" json:"quantity"`        // 回收数量
-	CostDeduction float64   `gorm:"not null" json:"cost_deduction"`  // 成本扣减金额
+	CostDeduction int64     `gorm:"not null" json:"cost_deduction"` // 成本扣减金额（单位：分）
 	CreatedAt     time.Time `json:"created_at"`
 }
